# hr/src/main/openapi/hr-openapi.yaml
openapi: "3.0.3"
info:
  title: "kwops API"
  description: "kwops API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8083/api"
paths:
  /employees/{number}:
    get:
      tags:
        - employee-controller
      summary: "GET employees/{number}"
      operationId: "getByNumber"
      parameters:
        - name: "number"
          example: "123456"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailsDto"
  /employees:
    get:
      tags:
        - employee-controller
      summary: "GET employees"
      operationId: "getAllEmployees"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllEmployees"
    post:
      tags:
        - employee-controller
      summary: "POST employees"
      operationId: "add"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeCreateDto"
        required: true
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
  /employees/{number}/dismiss:
    post:
      tags:
        - employee-controller
      summary: "POST employees/{number}/dismiss"
      operationId: "dismiss"
      parameters:
        - name: "number"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "withNotice"
          in: "query"
          required: false
          schema:
            type: "boolean"
            default: "true"
      responses:
        "200":
          description: "OK"
components:
  schemas:
    EmployeeDetailsDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        number:
          type: "string"
          example: "123456"
        lastName:
          type: "string"
          example: "Doe"
        firstName:
          type: "string"
          example: "John"
        startDate:
          type: "string"
          format: "date"
          example: "2021-01-01"
        endDate:
          type: "string"
          format: "date"
          example: "2021-12-31"
    EmployeeCreateDto:
      type: "object"
      properties:
        lastName:
          type: "string"
          example: "Doe"
        firstName:
          type: "string"
          example: "John"
        startDate:
          type: "string"
          format: "date"
          example: "2021-01-01"
    AllEmployees:
      type: "object"
      properties:
        employees:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeDetailsDto'